src/releng_mapper/requirements_frozen.txt
aioamqp==0.10.0
aiohttp==2.3.10
alembic==0.9.9
amqp==2.2.2
async-timeout==1.4.0
attrs==17.4.0
backcall==0.1.0
blinker==1.4
boto==2.48.0
boto3==1.7.1
botocore==1.10.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
clickclick==1.2.2
codecov==2.0.15
connexion==1.3
cookies==2.2.1
coverage==4.5.1
coveralls==1.3.0
decorator==4.2.1
docopt==0.6.2
docutils==0.14
ecdsa==0.13
fancycompleter==0.8
flake8==3.5.0
flake8-coding==1.3.0
flake8-copyright==0.2.0
flake8-debugger==3.1.0
flake8-isort==2.5
flake8-mypy==17.8.0
flake8-quotes==0.14.1
Flask==0.12.2
Flask-Cache==0.13.1
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Migrate==2.1.1
flask-oidc==1.3.0
Flask-SQLAlchemy==2.3.2
flask-talisman==0.5.0
future==0.16.0
gunicorn==19.7.1
httplib2==0.11.3
idna==2.6
idna-ssl==1.0.1
inflection==0.3.1
inotify==0.2.9
ipdb==0.11
ipython==6.3.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kombu==4.1.0
Logbook==1.3.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mohawk==0.3.4
more-itertools==4.1.0
mozdef-client==1.0.11
mozilla-backend-common==1.0.0
mozilla-cli-common==1.0.0
multidict==4.1.0
mypy==0.580
nose==1.3.7
oauth2client==4.1.2
parso==0.1.1
pdbpp==0.9.2
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycryptodome==3.5.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.5.0
pytest-cov==2.5.1
python-dateutil==2.6.1
python-editor==1.0.3
python-hglib==2.5
python-jose==2.0.2
pytz==2018.3
PyYAML==3.12
raven==6.6.0
requests==2.18.4
requests-futures==0.9.7
responses==0.9.0
rsa==3.4.2
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
slugid==1.0.7
SQLAlchemy==1.2.6
structlog==18.1.0
swagger-spec-validator==2.1.0
taskcluster==2.1.3
testfixtures==6.0.0
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.4
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
wmctrl==0.3
yarl==1.1.1
